requirements.txt
# Core ML/DL
torch>=1.12.0
fair-esm>=2.0.0
xgboost>=1.7.0
scikit-learn>=1.3.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
biopython>=1.81

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Statistics
scipy>=1.10.0

# Interpretability
shap>=0.42.0

# Utilities
tqdm>=4.65.0

# Notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0
